use page.ts instead of content.js
