+++ extension/src/manifest.json
@@ -58,12 +58,6 @@
             "js": ["asbplayer.js"],
             "all_frames": true,
             "run_at": "document_start"
-        },
-        {
-            "matches": ["<all_urls>"],
-            "js": ["content.js"],
-            "all_frames": true,
-            "run_at": "document_start"
         }
     ],
     "commands": {
